// services/notification-service/src/modules/auth/permission-key.enum.ts
export const enum PermissionKey {
  ViewAnyMessage = 'ViewAnyMessage',
  CreateAnyMessage = 'CreateAnyMessage',
  UpdateOwnMessage = 'UpdateOwnMessage',
  UpdateAnyMessage = 'UpdateAnyMessage',
  DeleteAnyMessage = 'DeleteAnyMessage',
  DeleteOwnMessage = 'DeleteOwnMessage',

  ViewAnyMessagingChannel = 'ViewAnyMessagingChannel',
  ViewOwnMessagingChannel = 'ViewOwnMessagingChannel',
  CreateMessagingChannel = 'CreateMessagingChannel',
  UpdateAnyMessagingChannnel = 'UpdateAnyMessagingChannnel',
  UpdateOwnMessagingChannel = 'UpdateOwnMessagingChannel',
  DeleteAnyMessagingChannel = 'DeleteAnyMessagingChannel',
  DeleteOwnMessagingChannel = 'DeleteOwnMessagingChannel',

  ViewAnyMessageTemplate = 'ViewAnyMessageTemplate',
  ViewOwnMessageTemplate = 'ViewOwnMessageTemplate',
  CreateMessageTemplate = 'CreateMessageTemplate',
  UpdateAnyMessageTemplate = 'UpdateAnyMessageTemplate',
  UpdateOwnMessageTemplate = 'UpdateOwnMessageTemplate',
  DeleteAnyMessageTemplate = 'DeleteAnyMessageTemplate',
  DeleteOwnMessageTemplate = 'DeleteOwnMessageTemplate'

}